update location of neovim config, update link.sh
* move ~/.nvim to ~/.config/nvim
* move ~/.nvimrc to ~/.config/nvim/init.vim
* update install/link.sh to skip already existing files
* update install/link.sh to handle creating symlinks for files in newly created config/ directory

These changes are consistent with the latest neovim install, which moves away from ~/.nvim and ~/.nvimrc

+++ install/link.sh
@@ -2,10 +2,32 @@
 
 DOTFILES=$HOME/.dotfiles
 
-echo "creating symlinks"
+echo -e "\nCreating symlinks"
+echo "=============================="
 linkables=$( find -H "$DOTFILES" -maxdepth 3 -name '*.symlink' )
 for file in $linkables ; do
     target="$HOME/.$( basename $file ".symlink" )"
-    echo "creating symlink for $file"
-    ln -s $file $target
+    if [ -e $target ]; then
+        echo "~${target#$HOME} already exists... Skipping."
+    else
+        echo "Creating symlink for $file"
+        ln -s $file $target
+    fi
+done
+
+echo -e "\n\ninstalling to ~/.config"
+echo "=============================="
+if [ ! -d $HOME/.config ]; then
+    echo "Creating ~/.config"
+    mkdir -p $HOME/.config
+fi
+# configs=$( find -path "$DOTFILES/config.symlink" -maxdepth 1 )
+for config in $DOTFILES/config/*; do
+    target=$HOME/.config/$( basename $config )
+    if [ -e $target ]; then
+        echo "~${target#$HOME} already exists... Skipping."
+    else
+        echo "Creating symlink for $config"
+        ln -s $config $target
+    fi
 done
